requirements/base.txt
Django==3.0.7
django-environ==0.4.5
psycopg2-binary==2.8.4
Pillow==7.1.2
djangorestframework==3.11.0
